feat: Add custom "Interaction To Next View" metric for Flutter.
Interaction To Next View measures the time from a user action, like a tap, until the next view is visible. On native platforms, this is timed until `startView` is called. However, in Flutter `startView` is called as soon as a route changes, before the view is visible to the user.

So, instead, Flutter measures the time from a user interaction until the completion of a view's "first build" (the FBC metric) and sets this as a `custom INV` value.

Because this calculation is so different from the native platforms, Flutter disables native INV calculations.

refs: RUM-8227