fix: Fix random trace id generation on web
Certain packages for distributed tracing on web (gRPC and GraphQL) generate tracing headers in Dart. In those cases, the library tries to generate a random 32-bit id, which throws on web. Fix so that these random ids are 31-bits instead.

Add an web integration test to ensure these trace ids work on web.

refs: #744