feat: Support simple Text and RichText widgets.
Text widgets are supported, but RichText currently concatenates all the text into a single string, which uses the default style on the root RichText widget.

WidgetSpans will get captured properly if they are in the center of the text, but they will overlap text in the player for the time being.

refs: RUM-9554