feat: Support basic capture of text fields.
Capturing text fields currently captures both the InputDecorator widget and the EditableText widgets. For now, UnderlineInputDecorator is pretending it is a rounded rectangle, though we will change that when we get support either for custom paint or per-side borders.

Some bugs still exist with capture, including currently capturing placeholder text even when it shouldn't be visible, and borders overlapping text in bordered fields.

refs: RUM-10254