chore: Remove use of GITHUB_TOKEN from release process
Instead of using PATs for Github operations, we're going to depend on the Github command line tool to do any operations involving Github.

Currently the only Github related commands revolve around querying the Android and iOS release versions, and creating the Draft release through the `deployer` script.

Also improve some functionality in the release process, including being able to run the script from a separate directory (to enable development of the script in parallel to writing a release).

refs: RUM-10261