feat!: Change TraceContextInjection default to sampled.
To better align SDKs, we're changing the default behavior for trace context injection from `all` to `sampled`.

`TraceContextInjection.all` added trace sampling decisions to all requests to first party hosts, regardless of the sampling decision.

`TraceContextInjection.sampled` only adds sampling information to requests that have been sampled in for tracing.

This is labeled as a breaking change because it changes behavior of the SDK, but most clients will not see a discernible difference in the behaviors, and it will not push a major version bump of the SDK.

refs: RUM-7748