feat(sr): First pass at supporting Android image capture
Current support for Android image processing is very slow, but confined to a background thread on Android.  However, because we halt collection until individual images are processed, this will negatively affect replays with a large number of images.  We'll need to look for an async method of processing images or speed it up significantly on Android

refs: RUM-11125

# Conflicts:
#	packages/datadog_session_replay/example/lib/app.dart
#	packages/datadog_session_replay/example/lib/routes.dart
#	packages/datadog_session_replay/example/lib/screens/main_screen.dart