chore: Make web integration tests more resilient
Instead of assuming requests in a certain order, we search for the correct request. This is because web can download more resources on a webpage than we expect, and we can't really know what order they'll come in.

Also, up memory request for Android CI machines, hopefully preventing gradle daemon from disappearing.