feat(sr): Support privacy options for text an images
This implements TextAndInputPrivacy and ImagePrivacy options both at a root capture level and through `SessionReplayPrivacy` override widgets.

Moved golden tests to `example` so that we could add and test asset privacy options in golden tests without polluting the package with unnecessary assets.

Touch input privacy is currently not implemented and will be a follow-up PR.

refs: RUM-11127 RUM-11126