feat: Improve capture performance
Instead of asking for semantics from every recorder, bind recorders to types so only the relevant ones are asked for capture semantics. This allows us to avoid capturing paint bounds and doing matrix translations for every element in the tree, only the ones that we need to capture semantics for.

Additionally, use `addPostFrameCallback` to make it so we only capture the tree if there was a change to the widget tree.

refs: RUM-11123