feat(sr): Implement touch privacy
The `SessionReplayPrivacy` widget can now override touch privacy for a tree. It does this by canceling touch events from a given pointer until the next capture. This can mean we might miss some taps if they are moving in and out of a private area within a capture period.

Add tests for both pointer recorder and associated widgets.

refs: RUM-11681