fix: GqlErrors causing stopResource ArgumentError
When handling gql response with errors, errors are serialized and passed to rum.stopResource as an attribute however, attributes must be of type Map<String, Object?> but when gqllink has an error an attribute of MappedListIterable<GraphQLError, Map<String, Object?>> is add. By calling toList() on the serialized errors stopResources will no longer throw an exception.