chore(ci): Fix a flake in SR.
There is a flake with recent changes in Session Replay that would cause the decorator field in EditableText to not be captured, but only when the size of the field was too small to fit it.

Require the text field to be at least of the height to support rendering the decorator.