feat: Support consistent sampling based on session id
Consistent trace sampling based on RUM Session ID allows RUM clients and the backend to come to a the same decision about sampling without relying on randomness for each trace.

This adds a listener for Session changes to RUM, and changes the `shouldSampleTrace` method to check the RUM session id first, before using the trace sampling if no session can be found.

refs: RUM-12492

# Conflicts:
#	packages/datadog_flutter_plugin/android/src/main/kotlin/com/datadoghq/flutter/DatadogRumPlugin.kt
#	packages/datadog_flutter_plugin/lib/src/rum/ddrum_method_channel.dart