fix: Fix issue re-initializing platform channel platform
Normally, the DatadogSdk should not be reinitialized, but the e2e tests are an exception.
Doing this reinitialization was causing an exception when calling `listen` multiple times on the isolate communication stream.

Make a minor change where the communication stream is initialized and listened to once as a member of the Datadog platform.

Also make a minor fix to have session broadcast pushed to the main thread.

# Conflicts:
#	packages/datadog_flutter_plugin/ios/datadog_flutter_plugin/Sources/DatadogSdkPlugin.swift