feat: Support getting context from background isolates
Instead of relying on Dart held cached context, utilize native platforms shared memory instead.

On Android, this just means asking for more context during the request.

On iOS, I've moved the Context listener to being a singleton which will cache it. Fetching it requires a specialized C method to bridge Flutter and Swift.

refs: RUM-12874