chore: Check in code generated by build_runner
While this code can be considered part of the "build process", Flutter it's not part of `flutter run` and is not run as part of pub deployments, only when `build_runner build` is run manually.

As a result, when changes are made to these generated files, developers have to remember to manually run `build_runner` in order to accept the changes, and forgetting can lead to at best an annoying error attempting to run an example app or, at worst, a broken deployment.

Though there's still debate, consensus is now that `.g.dart` files should be committed, and most Flutter developed packages do so.